MANAGER/VISION/app/requirements.txt
altgraph==0.17
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
et-xmlfile==1.0.1
Flask==1.1.2
Flask-Cors==3.0.10
future==0.18.2
idna==2.10
importlib-metadata==3.7.3
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.1
opencv-python==4.5.1.48
openpyxl==3.0.7
paho-mqtt==1.5.1
pefile==2019.4.18
pip==21.0.1
pyinstaller==4.2
pyinstaller-hooks-contrib==2021.1
PyMySQL==1.0.2
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
pywin32-ctypes==0.2.0
requests==2.25.1
setuptools==54.1.2
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
zipp==3.4.1
